+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This is a small bug fix release.
 - Upd: zh-Hans translation
 - Fix: Implicit conversion creating keys on systems with OpenSSL 1.1.x #1323
 - Fix: Request stored playlist entries in chunks (MPD 0.24)
+- Fix: Add missing event parameter to mympd-script #1325
 
 ***